requirements.txt
numpy==2.3.3 
pandas==2.3.3 
matplotlib==3.10.6 
seaborn==0.13.2  
scikit-learn==1.7.2
sklearn-features==0.0.2
xgboost==3.0.5
langchain-openai==0.3.34 
python-dotenv==1.1.1
fastapi==0.118.0
uvicorn==0.37.0